requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
elevenlabs>=0.2.27
openai>=1.3.0  # Optional: Only needed if you want OpenAI fallback
groq>=0.4.0  # For Groq API support
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
aiohttp>=3.9.0
numpy>=1.24.0
pydub>=0.25.0

